feat: implement client-side context compaction for Bedrock Converse path
Bedrock Converse API does not support server-side compaction (InvokeModel-only),
causing Claude Code loops to degrade when context grows large. This adds
transparent client-side compaction: accept compact_20260112 config, strip
thinking blocks from old messages when threshold exceeded, and skip compaction
content blocks gracefully instead of throwing.

Also includes: token usage breakdown display, API key quota tracking,
dashboard analytics improvements, and frontend usage record enhancements.

+++ backend/src/main/java/com/api2api/domain/gateway/service/DefaultGatewayInvocationService.java
@@ -9,6 +9,7 @@ import com.api2api.domain.protocol.model.UnifiedTokenUsage;
 import com.api2api.domain.routing.model.RouteCandidate;
 import com.api2api.domain.routing.model.RoutePlan;
 import java.time.Instant;
+import java.math.BigDecimal;
 import java.util.Objects;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +20,7 @@ public class DefaultGatewayInvocationService implements GatewayInvocationService
     public GatewayInvocation authenticate(
             GatewayInvocation invocation,
             ApiCredential credential,
-            long currentConsumedTokens,
+            BigDecimal currentConsumedTokens,
             Instant now
     ) {
         Objects.requireNonNull(invocation, "Gateway invocation must not be null");
