+++ backend/src/main/java/com/api2api/ohs/http/dashboard/dto/ChannelTokenTrendPointResponse.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 /**
  * Channel token trend bucket for admin dashboard.
@@ -19,5 +20,5 @@ public class ChannelTokenTrendPointResponse {
     private Instant bucketEnd;
     private Long providerChannelId;
     private String providerChannelName;
-    private long totalTokens;
+    private BigDecimal totalTokens;
 }
