+++ backend/src/main/java/com/api2api/ohs/http/dashboard/dto/UserTokenRankingResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 /**
  * User token ranking row for admin dashboard.
@@ -17,5 +18,5 @@ public class UserTokenRankingResponse {
     private int rank;
     private Long userAccountId;
     private String username;
-    private long totalTokens;
+    private BigDecimal totalTokens;
 }
